refactor(cli): extract commands/skills.ts and commands/doctor.ts (batch 3)
Move the two zero-reverse-dependency command modules out of qmd.ts,
per spec §11 (batch 3 first among command modules — earliest exposure
of import-graph issues, smallest rollback surface):

- commands/skills.ts: skill discovery, skills list/get/path, skill
  show/install + Claude symlink handling. Self-contained (no db).
- commands/doctor.ts: qmd doctor + qmd status with all check*/device
  helpers module-private. formatCount/shortModelName exported for
  vectorIndex (embed progress); showStatus/showDoctor exported for
  the dispatch switch.
- outputSkillsJson exported from skills.ts for the switch's --json
  error path.

Verified: tsc clean, build ok, 32 snapshots byte-identical,
cli/lazy-llm-import/exit-lifecycle green, cli.test.ts 146/146,
skills list + doctor smoke-tested.
